pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gqe-for-qsci"
version = "0.1.0"
description = "Training/evaluation pipeline for GQE (Generative Quantum Eigensolver) + QSCI (Quantum Selective Configuration Interaction) on molecular Hamiltonians"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Ryota Kemmoku"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Chemistry",
]
dependencies = [
    "torch==2.9.1",
    "pytorch-lightning==2.2.5",
    "scipy~=1.15.3",
    "hydra-core==1.3.2",
    "omegaconf==2.3.0",
    "wandb>=0.23.1",
    "tequila-basic==1.9.9",
    "pyscf==2.11.0",
    "ipykernel==6.30.0",
    "qc-PyCI==0.6.3",
]

[tool.hatch.build.targets.wheel]
packages = ["gqe_qsci"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]